Finish Studio 16

+++ SharedCode/TextFile.cpp
@@ -9,7 +9,7 @@ TextFile::TextFile(std::string s) : fileName(s) {
 };
 
 unsigned int TextFile::getSize() {
-	return content.size();
+	return static_cast<int>(content.size());
 }
 
 std::string TextFile::getName() {
